pyproject.toml
[tool.poetry]
name = "cabin-dog"
version = "0.1.0"
description = ""
authors = ["kenny-skaggs <skaggskd88@gmail.com>"]
readme = "README.md"
packages = [{include = "cabin_dog"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
djangorestframework = "^3.14.0"
faker = "^17.0.0"
python-dotenv = "^0.21.1"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
django-cors-headers = "^3.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
